+++ src/main/java/com/effectivemobile/testtask/bank/service/PhoneServiceImpl.java
@@ -1,6 +1,8 @@
 package com.effectivemobile.testtask.bank.service;
 
-import com.effectivemobile.testtask.bank.dto.AddPhoneDto;
+import com.effectivemobile.testtask.bank.dto.PhoneAddDto;
+import com.effectivemobile.testtask.bank.dto.PhoneChangeDto;
+import com.effectivemobile.testtask.bank.dto.PhoneDeleteDto;
 import com.effectivemobile.testtask.bank.dto.UserReturnDto;
 import com.effectivemobile.testtask.bank.mapper.UserMapper;
 import com.effectivemobile.testtask.bank.model.Phone;
@@ -8,7 +10,6 @@ import com.effectivemobile.testtask.bank.model.User;
 import com.effectivemobile.testtask.bank.repository.PhoneRepository;
 import com.effectivemobile.testtask.bank.repository.UserRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,27 +25,50 @@ public class PhoneServiceImpl implements PhoneService {
 
     @Transactional
     @Override
-    public UserReturnDto addPhone(String username, AddPhoneDto addPhoneDto) {
-        User user = userRepository.findByUserName(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        Phone phone = new Phone(addPhoneDto.getPhone());
+    public UserReturnDto addPhone(String userName, PhoneAddDto phoneAddDto) {
+        User user = checkUser(userName);
+        Phone phone = new Phone(phoneAddDto.getNumber());
         phone.setUser(user);
-        user.getPhones().add(phone);
-        return userMapper.toUserReturnDto(userRepository.save(user));
+        phoneRepository.save(phone);
+        return userMapper.toUserReturnDto(user);
     }
 
     @Transactional
     @Override
-    public void deletePhone(String username, String phone) {
-        User user = userRepository.findByUserName(username)
-                .orElseThrow(() -> new IllegalArgumentException("User not found"));
+    public void deletePhone(String userName, PhoneDeleteDto phoneDeleteDto) {
+        User user = checkUser(userName);
+        Phone deletePhone = new Phone(phoneDeleteDto.getNumber());
         Set<Phone> phones = user.getPhones();
         if (phones.size() < 2) {
             throw new IllegalArgumentException("Cannot delete the last phone");
         }
-        Phone deletePhone = phones.stream()
-                .filter(p -> p.getPhone().equals(phone))
-                .findAny().orElseThrow(() -> new IllegalArgumentException("Nothing to delete"));
+        if (!phones.contains(deletePhone)) {
+            throw new IllegalArgumentException("Nothing to delete");
+        }
+        phones.remove(deletePhone);
         phoneRepository.delete(deletePhone);
     }
+
+    @Transactional
+    @Override
+    public UserReturnDto changePhone(String userName, PhoneChangeDto phoneChangeDto) {
+        User user = checkUser(userName);
+        Phone oldPhone = new Phone(phoneChangeDto.getOldPhoneNumber());
+        Phone newPhone = new Phone(phoneChangeDto.getNewPhoneNumber());
+        Set<Phone> phones = user.getPhones();
+        if (!phones.remove(oldPhone)) {
+            throw new IllegalArgumentException("Phone to change not found");
+        }
+        newPhone.setUser(user);
+        phones.add(newPhone);
+        phoneRepository.save(newPhone);
+        return userMapper.toUserReturnDto(user);
+
+    }
+
+    private User checkUser(String username) {
+        return userRepository.findByUserName(username)
+                .orElseThrow(() -> new IllegalArgumentException("User not found"));
+    }
+
 }
